+++ attributes/default.rb
@@ -1,7 +1,7 @@
 # config file attributes
 default['remote_syslog']['conf']['files'] = []
 default['remote_syslog']['destination']['host'] = 'logs.papertrailapp.com'
-default['remote_syslog']['destination']['port'] = '12345'
+default['remote_syslog']['destination']['port'] = 12345
 default['remote_syslog']['hostname'] = node['hostname']
 default['remote_syslog']['exclude_files'] = []
 default['remote_syslog']['parse_fields'] = nil